requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
et_xmlfile==2.0.0
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
matplotlib==3.9.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
playwright==1.54.0
playwright-stealth==2.0.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.7
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0